requirements.txt
fastapi
uvicorn[standard]
requests
pyjwt
kafka-python
scikit-learn
